Moved string-checking-keywords.ts
