+++ gui/src/domain/string-checking-keywords.ts
